[JSC] `StringValueOf` node should not be converted to `ToString` node when `this` value is not string https://bugs.webkit.org/show_bug.cgi?id=284614
Reviewed by NOBODY (OOPS!).

The `ToString` node converts `undefined` into the string `"undefined"`.
However, `String.prototype.toString` and `String.prototype.valueOf`
throw a `TypeError` if the `this` value is not a string[1].

Therefore, converting a `StringValueOf` node into a `ToString` node is
incorrect.

This patch updates `DFGFixupPhase` to prevent the `StringValueOf` node
from being converted into a `ToString` node.

Also, there are no performance regressions in the benchmark added by
the patch that introduced the removed branch[2]:

                             TipOfTree                  Patched

string-or-other-add       88.0114+-0.7678           87.7158+-0.8438

[1]: https://tc39.es/ecma262/#sec-string.prototype.tostring
[2]: https://commits.webkit.org/275280@main

* Source/JavaScriptCore/dfg/DFGFixupPhase.cpp:
(JSC::DFG::FixupPhase::fixupStringValueOf):