[JSC] `RegExp#[Symbol.search]` should throw `TypeError` when `lastIndex` isn't writable https://bugs.webkit.org/show_bug.cgi?id=288832
Reviewed by NOBODY (OOPS!).

According to the spec[1], `RegExp#[Symbol.search]` writes to the `lastIndex`
property of `this` in the following cases:

- `lastIndex` is not 0
- The value of `lastIndex` changes before and after calling `RegExpExec`
  (which occurs when the `g` or `y` flag is enabled)

So, if `lastIndex` is not writable and the above conditions are met, a
`TypeError` should be thrown.

However, in the current JSC implementation, no error is thrown. This
patch changes to execute the slow path when `lastIndex` is not writable,
ensuring that an error is thrown.

                             TipOfTree                  Patched

regexp-search-basic        4.4263+-0.5470            4.3608+-0.2899          might be 1.0150x faster

[1]: https://tc39.es/ecma262/#sec-regexp.prototype-%symbol.search%

* Source/JavaScriptCore/builtins/BuiltinNames.h:
* Source/JavaScriptCore/builtins/RegExpPrototype.js:
(overriddenName.string_appeared_here.search):
* Source/JavaScriptCore/bytecode/LinkTimeConstant.h:
* Source/JavaScriptCore/runtime/JSGlobalObject.cpp:
(JSC::JSGlobalObject::init):
* Source/JavaScriptCore/runtime/RegExpObject.h:
* Source/JavaScriptCore/runtime/RegExpPrototype.cpp:
(JSC::JSC_DEFINE_HOST_FUNCTION):
* Source/JavaScriptCore/runtime/RegExpPrototype.h: